add libpolyhedron_demo
