+++ Three/doc/Three/Doxyfile.in
@@ -6,6 +6,7 @@ INPUT        = ${CMAKE_SOURCE_DIR}/Three/doc/Three/ \
                ${CMAKE_SOURCE_DIR}/Polyhedron/demo/Polyhedron/Viewer.h \
                ${CMAKE_SOURCE_DIR}/Polyhedron/demo/Polyhedron/Scene_item.h \
                ${CMAKE_SOURCE_DIR}/Polyhedron/demo/Polyhedron/MainWindow.h \
+               ${CMAKE_SOURCE_DIR}/Polyhedron/demo/Polyhedron/Polyhedron_demo.h \
                
 EXAMPLE_PATH = ${CMAKE_SOURCE_DIR}/Polyhedron/demo/Polyhedron/ \
